refactor: auth refactoring - dto class rename DTO to Dto - UserService rename
