+++ src/main/java/goorm/mybatisboard/auth/dto/LoginDto.java
@@ -6,8 +6,7 @@ import jakarta.validation.constraints.Size;
 import lombok.Data;
 
 @Data
-public class SignupDTO {
-
+public class LoginDto {
     @Email(message = "{validation.email.invalid}")
     @NotBlank(message = "{validation.email.required}")
     private String email;
@@ -19,7 +18,4 @@ public class SignupDTO {
     @Size(min=4, message = "{validation.password.size}")
     @NotBlank(message = "{validation.password.required}")
     private String password;
-
-    @NotBlank(message = "{validation.password.confirm.required}")
-    private String passwordConfirm;
 }
